requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
langchain>=0.2.0,<0.4.0
langchain-google-genai==2.0.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
langchain-community>=0.2.0,<0.4.0
google-generativeai>=0.5.0
PyPDF2==3.0.1
pydantic>=2.0.0,<3
aiohttp==3.8.5
